Move NSNotificationCenter observers to ViewDidDisappear/Appear.
This is supposed to make the app more stable, as dealloc
isn't always called.  Some views have been left alone, ie.
UserOverview, so vigilance is required to see if they are the
culprit for crashes.

Also, no 'success' message will be created after this change,
so another method of displaying the success is required.